ml-service/requirements.txt
tensorflow>=2.16.1
keras>=3.0.0
flask
flask-cors
pillow
opencv-python-headless
numpy
matplotlib
scikit-learn
requests
python-dotenv
